+++ src/pct/filters.c
@@ -1,4 +1,4 @@
-#include <filters.h>
+#include <pct/filters.h>
 #include <stdlib.h>
 #include <string.h>
