+++ src/pct/pthreads-conv.c
@@ -1,4 +1,4 @@
-#include <filters.h>
+#include <pct/filters.h>
 #include <pthread.h>
 #include <stdlib.h>
 #include <string.h>
